pulling from waypoints and requesting from yelp works

+++ src/main/java/finalproject/alongtheway/BusinessSearchApiService.java
@@ -47,7 +47,7 @@ public class BusinessSearchApiService {
 	}
 
 	// search with no filter by coords 
-	public List<Businesses> getAllResultsByCoord(Long latitude, Long longitude) {
+	public List<Businesses> getAllResultsByCoord(Double latitude, Double longitude) {
 		HttpHeaders headers = new HttpHeaders();
 		headers.add("Authorization", "Bearer " + yelpkey);
 		String url = "https://api.yelp.com/v3/businesses/search?longitude=" + longitude + "&latitude=" + latitude;
@@ -57,7 +57,7 @@ public class BusinessSearchApiService {
 	}
 	
 	// search by category given a coord set as longs
-	public List<Businesses> getAllResultsByCoordByCategory(Long longitude, Long latitude, String category){
+	public List<Businesses> getAllResultsByCoordByCategory(Double longitude, Double latitude, String category){
 		HttpHeaders headers = new HttpHeaders();
 		headers.add("Authorization", "Bearer " + yelpkey);
 		String url = "https://api.yelp.com/v3/businesses/search?longitude=" + longitude + "&latitude=" + latitude + "&=categories" + category;
