+++ src/main/java/finalproject/alongtheway/model/Center.java
@@ -2,19 +2,19 @@ package finalproject.alongtheway.model;
 
 public class Center {
 
-	private Long longitude;
-	private Long latitude;
+	private Double longitude;
+	private Double latitude;
 	
-	public Long getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(Long longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
-	public Long getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(Long latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
