+++ src/main/java/finalproject/alongtheway/model/Coordinates.java
@@ -2,30 +2,30 @@ package finalproject.alongtheway.model;
 
 public class Coordinates {
 
-	private Long latitude;
-	private Long longitude;
+	private Double latitude;
+	private Double longitude;
 
 	public Coordinates() {
 	}
 
-	public Coordinates(Long latitude, Long longitude) {
+	public Coordinates(Double latitude, Double longitude) {
 		setLatitude(latitude);
 		setLongitude(longitude);
 	}
 
-	public Long getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(Long latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 
-	public Long getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(Long longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
